requirements.txt
Django==1.9.7
django-recaptcha==1.0.5
